+++ Profiles.Persistence/Repositories/PatientRepository.cs
@@ -1,4 +1,6 @@
 ﻿using Dapper;
+using Events;
+using MassTransit;
 using Profiles.Domain.Entities;
 using Profiles.Domain.Entities.ForeignEntities;
 using Profiles.Domain.Interfaces.Repositories;
@@ -16,24 +18,24 @@ namespace Profiles.Persistence.Repositories
     public class PatientRepository : IPatientRepository
     {
         private readonly ProfilesDbContext _profileDbContext;
+        private readonly IPublishEndpoint _publishEndpoint;
 
-        public PatientRepository(ProfilesDbContext profilesDbContext)
+        public PatientRepository(ProfilesDbContext profilesDbContext, IPublishEndpoint publishEndpoint)
         {
             _profileDbContext = profilesDbContext;
+            _publishEndpoint = publishEndpoint;
         }
         public async Task<IEnumerable<Patient?>> GetAllPatients(CancellationToken cancellationToken)
         {
             var query = "SELECT * " +
                 "FROM [Patients] " +
-                "JOIN Accounts ON Patients.AccountId = Accounts.Id " +
-                "JOIN Photos ON Accounts.PhotoId = Photos.Id";
+                "JOIN Accounts ON Patients.AccountId = Accounts.Id";
 
             using (var connection = _profileDbContext.CreateConnection())
             {
-                var patients = await connection.QueryAsync<Patient, Account, Photo, Patient>(query,
-                    (patient, account, photo) =>
+                var patients = await connection.QueryAsync<Patient, Account, Patient>(query,
+                    (patient, account) =>
                     {
-                        account.Photo = photo;
                         patient.Account = account;
                         return patient;
                     });
@@ -46,7 +48,6 @@ namespace Profiles.Persistence.Repositories
             var query = "SELECT * " +
                 "FROM [Patients] " +
                 "JOIN Accounts ON Patients.AccountId = Accounts.Id " +
-                "JOIN Photos ON Accounts.PhotoId = Photos.Id " +
                 "WHERE Patients.Id = @Id";
 
             var parameters = new DynamicParameters();
@@ -54,10 +55,9 @@ namespace Profiles.Persistence.Repositories
 
             using (var connection = _profileDbContext.CreateConnection())
             {
-                var patients = await connection.QueryAsync<Patient, Account, Photo, Patient>(query,
-                    (patient, account, photo) =>
+                var patients = await connection.QueryAsync<Patient, Account, Patient>(query,
+                    (patient, account) =>
                     {
-                        account.Photo = photo;
                         patient.Account = account;
                         return patient;
                     }, param: parameters);
@@ -69,6 +69,7 @@ namespace Profiles.Persistence.Repositories
         {
             var query = "INSERT into [Patients] " +
                 "(FirstName, LastName, MiddleName, DateOfBirth, AccountId) " +
+                "OUTPUT INSERTED.Id " +
                 "values (@FirstName, @LastName, @MiddleName, @DateOfBirth, @AccountId)";
 
             var parameters = new DynamicParameters();
@@ -80,7 +81,20 @@ namespace Profiles.Persistence.Repositories
 
             using (var connection = _profileDbContext.CreateConnection())
             {
-                var r = await connection.ExecuteAsync(query, parameters);
+                var r = connection.ExecuteScalar<int>(query, parameters);
+
+                var insertedPatient = await GetPatient(r, default(CancellationToken));
+
+                await _publishEndpoint.Publish<PatientCreated>(new
+                {
+                    Id = r,
+                    FirstName = insertedPatient.FirstName,
+                    LastName = insertedPatient.LastName,
+                    MiddleName = insertedPatient.MiddleName,
+                    PhotoUrl = insertedPatient.Account.PhotoUrl,
+                    DateOfBirth = insertedPatient.DateOfBirth
+                });
+
                 return r == 0 ? null : patient;
             }
         }
@@ -103,19 +117,53 @@ namespace Profiles.Persistence.Repositories
             using (var connection = _profileDbContext.CreateConnection())
             {
                 var r = await connection.ExecuteAsync(query, parameters);
+
+                if (r != 0)
+                {
+                    var updatedPatient = await GetPatient(patient.Id, default(CancellationToken));
+
+                    await _publishEndpoint.Publish<PatientUpdated>(new
+                    {
+                        Id = updatedPatient.Id,
+                        FirstName = updatedPatient.FirstName,
+                        LastName = updatedPatient.LastName,
+                        MiddleName = updatedPatient.MiddleName,
+                        PhotoUrl = updatedPatient.Account.PhotoUrl,
+                        DateOfBirth = updatedPatient.DateOfBirth
+                    });
+                }
+
                 return r == 0 ? null : patient;
             }
         }
 
         public async Task<int> DeletePatient(int id, CancellationToken cancellationToken)
         {
-            string query = "DELETE FROM [Patients] WHERE Id = @Id";
-
-            using (var connection = _profileDbContext.CreateConnection())
+            var patient = await GetPatient(id, default(CancellationToken));
+            if (patient != null)
             {
-                var r = await connection.ExecuteAsync(query, new { Id = id });
-                return r;
+                string query = "DELETE FROM [Patients] WHERE Id = @Id";
+
+                using (var connection = _profileDbContext.CreateConnection())
+                {
+                    var r = await connection.ExecuteAsync(query, new { Id = id });
+
+                    if (r != 0)
+                    {
+                        await _publishEndpoint.Publish<PatientDeleted>(new
+                        {
+                            Id = patient.Id,
+                            FirstName = patient.FirstName,
+                            LastName = patient.LastName,
+                            MiddleName = patient.MiddleName,
+                            PhotoUrl = patient.Account.PhotoUrl,
+                            DateOfBirth = patient.DateOfBirth
+                        });
+                    }
+                    return r;
+                }
             }
+            return 0;
         }
     }
 }
